Moving API calls to Frontend
